Restructure context module
